Change Generate Repeat Event to singleton pattern
Implemented a singleton class Generate Repeat Event so we can use it
in edit command if needed.
