+++ src/main/java/seedu/scheduler/model/event/Event.java
@@ -2,15 +2,8 @@ package seedu.scheduler.model.event;
 
 import static seedu.scheduler.commons.util.CollectionUtil.requireAllNonNull;
 
-import java.time.DateTimeException;
-import java.time.Duration;
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoField;
-import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
@@ -103,156 +96,6 @@ public class Event {
      */
     public static boolean isValidEventDateTime(DateTime startDateTime, DateTime endDateTime) {
         return startDateTime.compareTo(endDateTime) <= 0;
-    }
-
-    /**
-     * Generate all repeated events from {@code targetEvent} according to its repeat type.
-     * {@code targetEvent} must have a valid event repeat type.
-     * Returns an unmodifiable list of repeated events.
-     */
-    public static List<Event> generateAllRepeatedEvents(Event targetEvent) {
-        switch(targetEvent.getRepeatType()) {
-        case DAILY:
-            return Collections.unmodifiableList(generateDailyRepeatEvents(targetEvent));
-        case WEEKLY:
-            return Collections.unmodifiableList(generateWeeklyRepeatEvents(targetEvent));
-        case MONTHLY:
-            return Collections.unmodifiableList(generateMonthlyRepeatEvents(targetEvent));
-        case YEARLY:
-            return Collections.unmodifiableList(generateYearlyRepeatEvents(targetEvent));
-        default:
-            return List.of(targetEvent);
-        }
-    }
-
-    /**
-     * Generate all events that are repeated daily from {@code targetEvent}.
-     * Returns a list of events that are repeated daily.
-     */
-    protected static List<Event> generateDailyRepeatEvents(Event targetEvent) {
-        List<Event> repeatedEventList = new ArrayList<>();
-        LocalDateTime repeatStartDateTime = targetEvent.getStartDateTime().value;
-        LocalDateTime repeatUntilDateTime = targetEvent.getRepeatUntilDateTime().value;
-        Duration durationDiff = Duration.between(targetEvent.getStartDateTime().value,
-                targetEvent.getEndDateTime().value);
-        while (repeatStartDateTime.isBefore(repeatUntilDateTime)) {
-            repeatedEventList.add(new Event(
-                    targetEvent.getUuid(),
-                    targetEvent.getEventName(),
-                    new DateTime(repeatStartDateTime),
-                    new DateTime(repeatStartDateTime.plus(durationDiff)),
-                    targetEvent.getDescription(),
-                    targetEvent.getVenue(),
-                    targetEvent.getRepeatType(),
-                    targetEvent.getRepeatUntilDateTime(),
-                    targetEvent.getTags()
-            ));
-            repeatStartDateTime = repeatStartDateTime.plusDays(1);
-        }
-        return repeatedEventList;
-    }
-
-    /**
-     * Generate all events that are repeated weekly from {@code targetEvent}.
-     * Returns a list of events that are repeated weekly.
-     */
-    protected static List<Event> generateWeeklyRepeatEvents(Event targetEvent) {
-        List<Event> repeatedEventList = new ArrayList<>();
-        LocalDateTime repeatStartDateTime = targetEvent.getStartDateTime().value;
-        LocalDateTime repeatUntilDateTime = targetEvent.getRepeatUntilDateTime().value;
-        Duration durationDiff = Duration.between(targetEvent.getStartDateTime().value,
-                targetEvent.getEndDateTime().value);
-        while (repeatStartDateTime.isBefore(repeatUntilDateTime)) {
-            repeatedEventList.add(new Event(
-                    targetEvent.getUuid(),
-                    targetEvent.getEventName(),
-                    new DateTime(repeatStartDateTime),
-                    new DateTime(repeatStartDateTime.plus(durationDiff)),
-                    targetEvent.getDescription(),
-                    targetEvent.getVenue(),
-                    targetEvent.getRepeatType(),
-                    targetEvent.getRepeatUntilDateTime(),
-                    targetEvent.getTags()
-            ));
-            repeatStartDateTime = repeatStartDateTime.plusWeeks(1);
-        }
-        return repeatedEventList;
-    }
-
-    /**
-     * Generate all events that are repeated monthly from {@code targetEvent}.
-     * Returns a list of events that are repeated monthly.
-     */
-    protected static List<Event> generateMonthlyRepeatEvents(Event targetEvent) {
-        List<Event> repeatedEventList = new ArrayList<>();
-        LocalDateTime repeatStartDateTime = targetEvent.getStartDateTime().value;
-        LocalDateTime repeatUntilDateTime = targetEvent.getRepeatUntilDateTime().value;
-        Duration durationDiff = Duration.between(targetEvent.getStartDateTime().value,
-                targetEvent.getEndDateTime().value);
-        while (repeatStartDateTime.isBefore(repeatUntilDateTime)) {
-            repeatedEventList.add(new Event(
-                    targetEvent.getUuid(),
-                    targetEvent.getEventName(),
-                    new DateTime(repeatStartDateTime),
-                    new DateTime(repeatStartDateTime.plus(durationDiff)),
-                    targetEvent.getDescription(),
-                    targetEvent.getVenue(),
-                    targetEvent.getRepeatType(),
-                    targetEvent.getRepeatUntilDateTime(),
-                    targetEvent.getTags()
-            ));
-            repeatStartDateTime = repeatStartDateTime.with((temporal) -> {
-                do {
-                    try {
-                        temporal = temporal.plus(1, ChronoUnit.MONTHS).with(ChronoField.DAY_OF_MONTH,
-                                targetEvent.getStartDateTime().value.getDayOfMonth());
-                    } catch (DateTimeException e) {
-                        temporal = temporal.plus(1, ChronoUnit.MONTHS);
-                    }
-                } while (temporal.get(ChronoField.DAY_OF_MONTH)
-                        != targetEvent.getStartDateTime().value.getDayOfMonth());
-                return temporal;
-            });
-        }
-        return repeatedEventList;
-    }
-
-    /**
-     * Generate all events that are repeated yearly from {@code targetEvent}.
-     * Returns a list of events that are repeated yearly.
-     */
-    protected static List<Event> generateYearlyRepeatEvents(Event targetEvent) {
-        List<Event> repeatedEventList = new ArrayList<>();
-        LocalDateTime repeatStartDateTime = targetEvent.getStartDateTime().value;
-        LocalDateTime repeatUntilDateTime = targetEvent.getRepeatUntilDateTime().value;
-        Duration durationDiff = Duration.between(targetEvent.getStartDateTime().value,
-                targetEvent.getEndDateTime().value);
-        while (repeatStartDateTime.isBefore(repeatUntilDateTime)) {
-            repeatedEventList.add(new Event(
-                    targetEvent.getUuid(),
-                    targetEvent.getEventName(),
-                    new DateTime(repeatStartDateTime),
-                    new DateTime(repeatStartDateTime.plus(durationDiff)),
-                    targetEvent.getDescription(),
-                    targetEvent.getVenue(),
-                    targetEvent.getRepeatType(),
-                    targetEvent.getRepeatUntilDateTime(),
-                    targetEvent.getTags()
-            ));
-            repeatStartDateTime = repeatStartDateTime.with((temporal) -> {
-                do {
-                    try {
-                        temporal = temporal.plus(1, ChronoUnit.YEARS).with(ChronoField.DAY_OF_MONTH,
-                                targetEvent.getStartDateTime().value.getDayOfMonth());
-                    } catch (DateTimeException e) {
-                        temporal = temporal.plus(1, ChronoUnit.YEARS);
-                    }
-                } while (temporal.get(ChronoField.DAY_OF_MONTH)
-                        != targetEvent.getStartDateTime().value.getDayOfMonth());
-                return temporal;
-            });
-        }
-        return repeatedEventList;
     }
 
     /**
